Overhaul of random equipment.  The weights on random eq should now be more realistic, and the affects on them shouldn't be overwritten quite as much.  Added a few material types.  Cleaned up material_restrictions function.  Added the various OBJCRAFT_... macros.  Made the random eq tables look more like tables.  Fixed setsuffix_obj_new to only have possible cases in the switch statement.

+++ src/prototypes.h
@@ -1043,7 +1043,7 @@ void create_randoms();
 bool check_random_drop(P_char ch, P_char mob, bool piece);
 P_obj create_random_eq(int charlvl, int moblvl, int item_type, int material_type);
 P_obj create_random_eq_new(P_char killer, P_char mob, int item_type, int material_type);
-P_obj setprefix_obj(P_obj obj, int modifier, int affectnumber);
+P_obj setprefix_obj(P_obj obj, float modifier, int affectnumber);
 P_obj create_stones(P_char ch);
 P_obj create_material(int index);
 P_obj create_material(P_char ch, P_char mob);
